Token now stores userid

+++ Server/Services/TransferModels/Requests/CreatePlayerDTO.cs
@@ -2,6 +2,6 @@ namespace Services.TransferModels.Requests;
 
 public class CreatePlayerDTO
 { 
-        public decimal InitialBalance { get; set; } = 0;  
+        public int InitialBalance { get; set; } = 0;  
         public bool IsActive { get; set; } = false; 
 }
